master - Persisting Android rating

+++ src/main/kotlin/com/kotlinklub/core/services/AndroidRatingService.kt
@@ -4,7 +4,7 @@ import com.kotlinklub.core.domain.AndroidRating
 import java.util.*
 
 interface AndroidRatingService {
-    fun create(date: Date, rating: Int): AndroidRating
+    fun create(date: Date, rating: Double): AndroidRating
     fun find(id: Long): AndroidRating?
     fun findAll(): List<AndroidRating>
 }
